[#13] style : google login 관련 파일 정리
